IntType& parameter renamed to IntType to allow for both reference and copy variable usage of parameter
The commit message summarizes the code changes made in the AltNum/AltDec.hpp file. The main change is that the IntType& parameter in several functions has been renamed to IntType, allowing for both reference and copy variable usage of the parameter. This change improves flexibility and usability of these functions.