negative status check comparison fix
Negative status check comparison fix

The commit fixes a bug in the code where the negative status check comparison was not working correctly. The variable "WasNegative" is introduced to store the previous negative status, and then used to compare with the current negative status. If there is a change in the negative status, the decimals are inverted. This fix ensures that the correct comparison is made and the decimals are inverted when necessary.

Additionally, the preprocessor toggle section has been separated into it's own region.

These changes improve the accuracy of negative status checks and provide region to toggle section of code.